+++ src/sonarcloud-api/issues/search/params/additional-field.ts
@@ -1,0 +1,11 @@
+export enum AdditionalField {
+  All = "_all",
+  Comments = "comments",
+  Languages = "languages",
+  ActionPlans = "actionPlans",
+  Rules = "rules",
+  RuleDescriptionContextKey = "ruleDescriptionContextKey",
+  Transitions = "transitions",
+  Actions = "actions",
+  Users = "users",
+}