+++ src/sonarcloud-api/issues/search/params/resolution.ts
@@ -1,0 +1,6 @@
+export enum Resolution {
+  FalsePositive = "FALSE-POSITIVE",
+  WontFix = "WONTFIX",
+  Fixed = "FIXED",
+  Removed = "REMOVED",
+}