+++ src/sonarcloud-api/issues/search/params/sort-field.ts
@@ -1,0 +1,10 @@
+export enum SortField {
+  FileLine = "FILE_LINE",
+  HotSpots = "HOTSPOTS",
+  CloseDate = "CLOSE_DATE",
+  UpdateDate = "UPDATE_DATE",
+  Status = "STATUS",
+  Assignee = "ASSIGNEE",
+  CreationDate = "CREATION_DATE",
+  Severity = "SEVERITY",
+}